Mouse detects tile base
Fixed decimal clipping (float to int)

+++ Motor2D/Map.cpp
@@ -99,7 +99,7 @@ std::pair<float, float> Map::F_WorldToMap(float x, float y) const
 	return map.F_WorldToMap(x, y);
 }
 
-std::pair<int, int> Map::WorldToTileBase(float x, float y) const
+std::pair<int, int> Map::WorldToTileBase(int x, int y) const
 {
-	return map.WorldToTileBase(x, y);
+	return map.WorldToTileBase(float(x), float(y));
 }
