+++ Motor2D/Map.h
@@ -29,7 +29,7 @@ public:
 	std::pair<float, float> F_MapToWorld(float x, float y) const;
 	std::pair<float, float> F_WorldToMap(float x, float y) const;
 
-	std::pair<int, int> WorldToTileBase(float x, float y) const;
+	std::pair<int, int> WorldToTileBase(int x, int y) const;
 
 private:
